fix: brands, category, search, menu, add productsList
