+++ src/Pages/BrandsPage/BrandsPage.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import Brands from "../../modules/components/Brands/Brands";
-import Category from "../../modules/components/Category/Category";
 import { useLocation } from "react-router-dom";
 import AllBrands from "../../modules/components/AllBrands/AllBrands";
