+++ src/modules/components/ProductsList/ProductsList.styled.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const ProductListContainer = styled.ul`
+  display: flex;
+  flex-wrap: wrap;
+  margin: -6px;
+`;
